LayoutTests/platform/efl/svg/dynamic-updates/SVGFilterElement-svgdom-filterResY-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'filterResY' property of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS filterElement.filterResY.baseVal is 400
PASS filterElement.filterResY.baseVal is 200
PASS successfullyParsed is true

TEST COMPLETE
